reducing lp norms

+++ models/shufflenet.py
@@ -29,7 +29,7 @@ class Bottleneck(nn.Module):
         self.lp_norm = lp_norm
         self.device = device
 
-        mid_planes = out_planes/4
+        mid_planes = int(out_planes/4)
         g = 1 if in_planes==24 else groups
         self.conv1 = nn.Conv2d(in_planes, mid_planes, kernel_size=1, groups=g, bias=False)
         # self.bn1 = nn.BatchNorm2d(mid_planes)
